polygon Connection (#74)
* Add CI script for Dev (#13)

* chore/CICD/12 (#14)

* Add CI script for Dev

* chore/12/CICD

* add build

* BuldTestReTry

* thirdTrialforBuildTest

* fourthTrialforTest

* fifthTrialforBuldTest

* Delete build directory

* Delete .github/workflows/dev-deploy.yml

* Feat/signin login/#15 (#18)

* Docs/SetSwagger#15/

* feat/UserSignup/#15

* fix/SignupBug/#16 (#19)

* Feat/for connect flask (#21)

* fix/SignupBug/#16

* feat/uploadVideo

* chore/modifyGradle

* feat/17/login (#22)

* fix/loginurl Security Setting

* feat/#25/uploadS3 (#28)

* feat/#25/uploadS3

* fix/#25/gitAction

* Update README.md

* Update README.md

* feat/#26/connectFlask (#29)

* feat/#26/connectFlask

* feat/#26/resultFromFlask

* Feat/#31/connectToTruffle1 (#32)

* feat/#31/connectToTruffle1

* fix/forBuildingGradle

* add gitignore

* Delete .DS_Store

* feat/modify Description on API

* fix/test code

* feat/#31/deploy (#37)

* Feat/#31/deploy (#39)

* feat/#31/deploy

* modify env setting

* Feat/#31/deploy (#40)

* feat/#31/deploy

* modify env setting

* modify script

* Feat/#31/deploy (#41)

* feat/#31/deploy

* modify env setting

* modify script

* modify2

* Feat/#31/deploy (#42)

* feat/#31/deploy

* modify env setting

* modify script

* modify2

* bugfix/dockercompose file

* chore/#43/optimization (#44)

* fix/#46,#24/fixing signup and login bug (#47)

* fix/#46,#24/fixing signup and login bug

* feat/#46/duplicationCheckEmail

* fix hbm2ddl.auto

* modify name of database

* chore/setCORS (#53)

* Chore/#52/set cors (#56)

* chore/setCORS

* change cors port number

* fix/#60/modifySecurity (#61)

* fix/#63/presignedUrl (#64)

* feat/#55/videoIsOpen (#65)

* Feat/#55/is open (#67)

* feat/#55/videoIsOpen

* fix/modifyGanache

* feat/#51/connectPolygon (#73)

* feat/#51/connectPolygon

* feat/#51,69/connctedWithPolygon

+++ src/main/java/TokenUs/TokenUs_BE/domain/Nft.java
@@ -1,5 +1,6 @@
 package TokenUs.TokenUs_BE.domain;
 
+import java.math.BigInteger;
 import jakarta.persistence.*;
 
 import lombok.*;
@@ -20,11 +21,11 @@ public class Nft extends BaseEntity {
     @Column(length = 500)
     private String contractAddress;
 
-    private Integer tokenId;
+    private BigInteger tokenId;
 
-    private Float currentPrice;
+    private BigInteger currentPrice;
 
-    private Integer mintQuantity;
+    private BigInteger mintQuantity;
 
     @Column(nullable = false, columnDefinition = "BOOLEAN DEFAULT false")
     private Boolean isListed;
